# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AYMANE Account API
  version: 1.0.0
  description: API interne de synchronisation du compte durable AYMANE.
servers:
  - url: /api
paths:
  /health:
    get:
      summary: Vérifier le service
      operationId: health
      responses:
        "200":
          description: Service disponible
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
  /account/me:
    get:
      summary: Lire le compte durable courant
      operationId: getAccount
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Compte courant
          content:
            application/json:
              schema:
                type: object
                required: [account]
                properties:
                  account:
                    $ref: "#/components/schemas/Account"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /account/sync:
    post:
      summary: Synchroniser la session avec le compte durable
      operationId: syncAccount
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventType]
              properties:
                eventType:
                  type: string
                  enum: [signup, signin, session]
      responses:
        "200":
          description: Compte synchronisé
          content:
            application/json:
              schema:
                type: object
                required: [account]
                properties:
                  account:
                    $ref: "#/components/schemas/Account"
                  welcomeMessage:
                    type: string
                    nullable: true
        "400":
          description: Requête invalide
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          description: Corps trop volumineux
        "429":
          $ref: "#/components/responses/RateLimited"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Account:
      type: object
      required: [id, email, role, emailVerified, createdAt, loginCount]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        fullName:
          type: string
          nullable: true
        role:
          type: string
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        loginCount:
          type: integer
          minimum: 0
  responses:
    Unauthorized:
      description: Session absente ou invalide
    RateLimited:
      description: Trop de requêtes